backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
httpx
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
authlib
alembic
